symfony 5 and customization/modernization; getting rid of doctrine
